+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
-## [unreleased]
+## [0.8.0] - 2023-07-29
 
 - Color .go filepaths inside the current working dir "cyan"
 - Color .go filepaths using "cyan;dim"